+++ src/modules/consumer/consumer.controller.ts
@@ -2,14 +2,14 @@ import { Body, Controller, Post } from '@nestjs/common'
 import { ApiBearerAuth, ApiTags } from '@nestjs/swagger'
 import { CurrentPrincipal } from '@wlisfes/chat-web-base-schema/auth'
 import type { AuthPrincipal } from '@wlisfes/chat-web-base-schema/auth'
-import { ConsumersService } from '@/modules/consumers/consumers.service'
-import { CreateConsumerDto, ListConsumerDto, UpdateConsumerDto, UpdateConsumerStatusDto } from '@/modules/consumers/dto/consumer.dto'
+import { ConsumerService } from '@/modules/consumer/consumer.service'
+import { CreateConsumerDto, ListConsumerDto, UpdateConsumerDto, UpdateConsumerStatusDto } from '@/modules/consumer/dto/consumer.dto'
 
 @ApiTags('账号中心-外部客户')
 @ApiBearerAuth('authorization')
-@Controller('consumers')
-export class ConsumersController {
-    constructor(private readonly service: ConsumersService) {}
+@Controller('consumer')
+export class ConsumerController {
+    constructor(private readonly service: ConsumerService) {}
 
     @Post('create')
     create(@CurrentPrincipal() principal: AuthPrincipal, @Body() input: CreateConsumerDto) {